Split functional tests into two - one for PostGres, one for Certificate Transparency.
